JavaScript-Load-Image
JavaScript Load Image is a library that allows you to load and manipulate images using JavaScript. It provides a simple, easy-to-use interface for working with images in the browser. It can be used to resize and crop images, adjust their orientation, and apply various filters and effects. It can also be used to load images from different sources, such as file inputs, URLs, and data URLs. The library is lightweight, fast, and easy to use, and has a number of advanced features, such as support for uploading images to the server and rendering images with canvas elements.